refactor: separate types
