+++ src/types/index.ts
@@ -1,39 +1,4 @@
-import type { EsDay } from '~/core/EsDay'
-
+export * from './core'
+export * from './date'
 export * from './locale'
-
-export type DateType = EsDay | Date | string | number | number[] | null | undefined | object
-export type PrettyUnitType = 'date' | 'day' | 'week' | 'month' | 'year'
-  | 'hour' | 'minute' | 'second' | 'millisecond'
-export type PrettyUnit<T extends UnitType> = T extends 'D' ? 'date' :
-  T extends 'w' ? 'week' : T extends 'M' ? 'month' : T extends 'y' ? 'year' :
-    T extends 'h' ? 'hour' : T extends 'm' ? 'minute' : T extends 's' ? 'second' :
-      T extends 'ms' ? 'millisecond' : T extends 'd' ? 'day' : T
-
-export type UnitType = 'y' | 'M' | 'D' | 'w' | 'h' | 'm' | 's' | 'ms' | 'd' | PrettyUnitType
-export interface AllDateFields {
-  year: number
-  month: number
-  /**
-   * date in month
-   * 1 ~ 31
-   */
-  date: number
-  /**
-   * day in week
-   * 0 ~ 6
-   */
-  day: number
-  hour: number
-  minute: number
-  second: number
-  millisecond: number
-}
-
-export type EsDayFactoryParserFn = (d?: DateType, ...others: any[]) => EsDay
-export interface EsDayFactory {
-  (...args: Parameters<EsDayFactoryParserFn>): ReturnType<EsDayFactoryParserFn>
-  utc: EsDayFactoryParserFn
-  extend: <T extends {}>(plugin: EsDayPlugin<T>, option?: T) => EsDayFactory
-}
-export type EsDayPlugin<T extends {}> = (option: T, dayTsClass: typeof EsDay, esday: EsDayFactory) => void
+export * from './unit'
